Возвращены старые перегрузки методов и помечены как Obsolete. ConsulMetricsHttpClientWrapper.cs использует асинхронные перегрузки ConsulServiceAddress.cs. Возвращено время обновления кеша в 5 секунд.
